Ajustes Generales Compras

+++ Core/DTO/COMPRAS/ProveedoresDTO.cs
@@ -8,7 +8,7 @@ namespace Core.DTO.COMPRAS
 {
     public class ProveedoresDTO
     {
-        public long noProveedor { get; set; }
+        public int noProveedor { get; set; }
         public string nomProveedor { get; set; }
         public int tipoCambio { get; set; }
         public string noProveedorStarsoft { get; set; }
